Crate paralegal

The paralegal annotation library.

For the comprehensive documentation see https://justus-adam.notion.site/Markers-and-Annotations-b3b078ea2f7f41739de1efe7f9d33484?pvs=4

Note that these proc macros expand differently depending on whether this library is compiled as part of paralegal analysis. As such you may encounter some compile errors at the use-site only when running cargo paralegal-flow.

Modules

impl_ 🔒

Macros

export 🔒

Re-export the named item from the impl_ mod as a proc-macro.

Attribute Macros

analyze

Mark this function as entry point for analysis.

marker

Attach an annotation to this item.

output_types

Declare an analysis-level type alias